Removing hard-coded WatchmakerUrl string.
Removed hard-coded url to Watchmaker egg since it contained words we don't want displayed. Need to pass in a url either by hard-coding it in when using in userdata for EC2 instance launches or by passing as param when using the script within PowerShell.

Added test for empty WatchmakerUrl - if empty, throws a log that Watchmaker is not installed. Otherwise, install Watchmaker.