Removes validation of saltstates from cli.py
saltstates has two special values, `None` and `Highstate`, both of
which are special-cased in the salt worker, and comma separated
strings are handled by salt-call directly. Attempting to validate
the special values and the comma separated string option basically
means the validation allows anything anyway, so the validation is
kind of pointless.

For now, any exception encountered will be handled by the salt
worker.

Fixes #56