Returns full path to formula directory
os.listdir only returns the name of the directory, not the full
path to the directory. The isdir() test was then failing, so the
method was always returning an empty list.

This patch uses `next(os.walk(salt_formula_root))[1]` to get all
the directories in the first level of salt_formula_root, which
again returns only the names of the directories. The list
comprehension joins them with salt_formula_root to return the full
path.

Fixes #180