Fixes error on EL6 with pip v7.1.0
The version of pip in epel for EL6 is v7.1.0, which thinks the
pypi.org index is an external link. That caused the install to
fail.

This patch changes the usage doc to install/upgrade pip first,
then to install watchmaker using the newer version of pip, which
does not have the same problem with the pypi.org index.