Parses state return as a python object
Prior attempt would still output all "changed" states in the
traceback. Salt does not appear to have an output option that
will output only failed states. Approach in this patch returns
the state output as a python object (formatted as a string),
converts it back to python using ast.literal_eval, and parses
the output for failed states.