Deploys watchmaker to GitHub releases on pushes to master
Updates travis-ci to deploy to GitHub releases after tests complete
successfully on the master branch. We read the version from setup.cfg,
set the tag, then travis uses the tag to create the release (which
also creates the tag in the GitHub repo).

Once the tag is pushed via the GH release, travis-ci runs again,
this time triggering the on.tags deploy to pypi.

The deploy stage in travis-ci will not run if the tests fail, and
will not execute on pull requests. This is expected to execute only
after a pull request is merged to the master branch, which we only
do for a release. (All other PRs are to the develop branch.)