Uses SERVER_AUTH for ssl context
Watchmaker has had this bug for a while now. CLIENT_AUTH is used by
servers to verify clients. SERVER_AUTH is used by clients to verify
servers.

Python 3.10 introduces PROTOCOL_TLS_CLIENT and PROTOCOL_TLS_SERVER,
which creates unique contexts for each scenario. Because Watchmaker
reversed the purpose for the context, it began to fail in Python 3.10.

For evidence, see the source for `create_default_context`:

* https://github.com/python/cpython/blob/b494f5935c92951e75597bfe1c8b1f3112fec270/Lib/ssl.py#L754-L760